# Convert escapeMetricName to use strings.Builder