@matthiasr 2019-05-17
TIL about strings.Builder 😄 thank you!